requirements-macos.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.12.5
pydantic-settings==2.1.0

opencv-python==4.9.0.80
numpy==1.26.4
Pillow==12.1.0
networkx==3.2.1
pytesseract==0.3.10
rapidfuzz==3.6.1

requests==2.32.3
jinja2==3.1.4
